requirements.txt
asgiref==3.7.2
Brotli==1.1.0
click==8.1.7
colorama==0.4.6
dj-database-url==2.2.0
Django==4.2.6
django-cors-headers==4.3.0
django-restframework==0.0.1
djangorestframework==3.14.0
gunicorn==22.0.0
h11==0.14.0
packaging==24.1
psycopg2-binary==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.12.2
tzdata==2023.3
uvicorn==0.30.1
whitenoise==6.7.0
